Update how content is displayed
